#159 #145 - 수정완료, FXKeychain버그로 인해 이런식을 처리

+++ I_NScreen/I_NScreen/Classes/_Menu/_HomeGnb/HomeGnbViewController.h
@@ -37,6 +37,6 @@
 @protocol HomeGnbViewDelegate <NSObject>
 
 @optional
-- (void)onHomeGnbViewMenuList:(int)nTag;
+- (void)onHomeGnbViewMenuList:(NSInteger)nTag;
 
 @end
